make docs more generic to package managers
for a long time NPM has basically been the de facto front end package manager. however, BunJS is making a strong push, and a good case for itself, so I think it'd be good for us to include it where appropriate.

we all know specific tooling comes and goes, so I've made the language *more generic* where appropriate, and refer to "a package manager" instead of specifically "NPM".

I've also then included the analogous `bun` command next to `npm` commands.