Modify the CSP Evaluator to handle negated expressions
This commit adds support to use straightforward negated expressions in
Alpine directives in the CSP build.

While we (probably) do not want to rewrite a full CSP-save expression
evaluator to handle expressions of arbitrary complexity, directives with
negation are such a common use case that it makes sense to support them.

Consider, for example,

* `x-show="!loading"`
* `:disabled="!valid"`

In the CSP, we would have to write custom getters for each boolean
attribute whose negation we might want to use, adding needless
verbosity.

On the other hand, being a unary operator, hand-rolling negation is easy
enough.

To test that it all works, we add relevant cypress tests to the csp
spec to verify that we can negate both a plain attribute as well as the
result of a function call.