update `x-cloak` CSS to allow disabling
change the suggested `x-cloak` CSS to specifically only target the boolean `x-cloak` attribute or the attribute with an empty string value (`x-cloak=""`).

this means we will no longer target the attribute when it has a non-empty string value (`x-cloak="off"`).

the benefit this gives is we can conditionally disable the cloaking for elements we know will be visible immediately. this prevents a "blip in" from the hidden to visible state.

Here is a Blade example to show how it would work:

```blade
<div
    x-show="jsCondition"
    x-cloak="{{ $condition ? 'off' : '' }}"
>Alpine Rocks</div>
```